# Command reorganization to reduce import overhead