--- a/se/cmd/help.py
+++ b/se/cmd/help.py
@@ -1,0 +1,21 @@
+"""
+help command
+"""
+
+from se.main import get_commands
+
+def cmd_help() -> int:
+	"""
+	Entry point for `se help`
+
+	help() is a built-in function so this function is called se_help().
+	"""
+
+	commands = get_commands()
+
+	print("The following commands are available:")
+
+	for command in commands:
+		print(command)
+
+	return 0