This is a pretty big change and perhaps a bit half-baked, so please consider this an exploration of some ideas rather than a change to be merged. And I saw in the other thread that you would like to do a release, so go ahead and do that. I don't have any other changes to go in.

After some thought, it seemed like a good way of providing custom imports for each command was to have an independent module for each command. So there are three changes included here:

1) The executables.py file is removed and the functions for each command have been moved to a command module under `se/cmd`. The entrypoint function for each command is `cmd_<name>()`.

2) The functions in `se/__init__.py` have been moved to `se/common.py` so that every command does not have to take the performance hit of the imports required by those functions. I'm not sure if this was really necessary. It is likely only saving 10-20ms for a few commands.

3) The entrypoint for the `se` command has moved to `se/main.py`. The implementation has changed slightly to read the command list from the list of modules under `se.cmd`.

On my machine the change in execution time (in seconds) is as follows:

command|old time|new time
-----------|---------|----------
build|6.930|6.830
lint|1.446|1.304
typogrify|1.383|1.225
dec2roman|0.389|0.080

So is it worth it? I'm on the fence. There is a pretty good performance improvement in the small commands but the most common commands have not improved much.

I kind of like the idea of having a module for each command. It makes it easy to find the code for a command and provides a natural place to put command-specific functions. I think it also makes it more obvious what the dependencies are for each command. But again, I'm not sure if the disruption to the status quo is worth the modest performance improvement.

Thoughts?