@acabal 2020-02-10
Interesting idea. It doesn't look like a huge win performance-wise; those differences are so small it may just be random. But it's interesting from an organizational perspective.

If we're putting the executable functions into a `cmd` module, then why preface them with `cmd_`?

We could also move `executables_create_draft.py` into `cmd`.

I would prefer to spell it out as `commands` since we are rarely using shorthand in the rest of the codebase.

So I suppose the question is, is this a major win for code organization? I guess that's a little open to interpretation. Would a Python expert find the new structure easier to navigate than the old? Or would either one do?