@hallidave 2020-02-12
The naming is pretty arbitrary.  It is no problem to change it as long as it's somewhat logical and consistent.

I'm not sure about Python experts, but I would say that being able to find the code for the typogrify command (as an example) in a module called `se.command.typogrify` is definitely going to make it easier for most people to find the code they're looking for.

And having the import list at the top of the `typogrify.py` file reflect all of the dependencies for that command (and only that command) makes the dependencies more obvious. Pruning the import list also has the nice side-effect of making some commands run faster, which was the original thing that I was exploring by reorganizing the code this way.

Anyways, if you think this is a interesting idea I can submit a pull request with some revised naming. Or we could do some sort of hybrid where some commands get their own files and others stay grouped together. Or we can just drop it. You won't hurt my feelings if you say "nah, not interested." :-)