@acabal 2020-02-13
I gave it some thought and I think it would be helpful to reorganize in this way. I think you're right that it's clearer which command needs which import, and it will be clearer to see them listed in a folder instead of all in a single file. This way we can also get rid of the lonely create_draft.py file and put it with the other commands. Let's call the module `commands` (plural).