@hallidave 2020-02-13
Sounds good. I will create a new pull request that just splits executables.py up into individual command files under a new `se.commands` module. Did you have an opinion on what to call the entry function for each command. Do we want `command_<name>()` or just `<name>()` or maybe `main()`?