@acabal 2020-02-13
I think `<name>()` will be good. IMHO `main()` should be reserved for the single entry point of the package, otherwise it gets confusing. Prefixing it with `command_` would be redundant because it's already in the `commands` module.