@acabal 2020-02-13
Can you wait till tomorrow to work on this? I'm tweaking the lint function executable right now and if you start work now then you'll have to deal with some merge nonsense.