@hallidave 2020-02-13
Sure. Not a problem.